Share unique map deserializer
