+++ crates/uv-toml/Cargo.toml
@@ -16,5 +16,6 @@ doctest = false
 workspace = true
 
 [dependencies]
+serde = { workspace = true }
 toml_datetime = { workspace = true }
 toml_parser = { workspace = true }
